% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{AFP}
\alias{AFP}
\title{Air filled porosity (AFP)}
\usage{
AFP(tp, gwc)
}
\arguments{
\item{tp}{total porosity}

\item{gwc}{gravimetric water content}
}
\value{
Air filled porosity (AFP)
}
\description{
Calculates the Air filled porosity (AFP)
}
\examples{
temp1 <- AFP(tp =49,gwc =15);
temp2 <- AFP(tp = c(48,44,49),gwc =c(15,20,10));
}
