% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{DS}
\alias{DS}
\title{Degree of saturation (DS)}
\usage{
DS(vwc, tp)
}
\arguments{
\item{vwc}{volumetric water content}

\item{tp}{total porosity}
}
\value{
Degree of saturation (DS)
}
\description{
Calculates the Degree of saturation (DS)
}
\examples{
temp1 <- DS(vwc= 19,tp=49.5);
temp2 <- DS(vwc = c(19,23,25),tp =c(48,46,49));
}
