% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{DSW}
\alias{DSW}
\title{Depth of soil water (DSW) (mm/cm or cm/m)}
\usage{
DSW(vwc, sd)
}
\arguments{
\item{vwc}{volumetric water content}

\item{sd}{soil depth (cm or m)}
}
\value{
Depth of soil water (DSW) (mm/cm or cm/m)
}
\description{
Calculates the Depth of soil water (DSW) (mm/cm or cm/m)
}
\examples{
temp1 <- DSW(vwc =23,sd =15);
temp2 <- DSW(vwc = c(25,36,29),sd =c(15,20,10));
}
