% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{GMD}
\alias{GMD}
\title{Geometric mean diameter (GMD)}
\usage{
GMD(WC, ADC)
}
\arguments{
\item{WC}{Weight of soil retained (g)}

\item{ADC}{average diameter between two sieves (mm)}
}
\value{
Geometric mean diameter (GMD), mm
#' @examples
temp1 <- GMD(WC = c(4.237,42.42,36.326,20.34),
ADC=c(5,1.1265,0.153,0.0265));
}
\description{
Calculates the Geometric mean diameter (GMD)
}
