% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{GWC}
\alias{GWC}
\title{Gravimetric water content}
\usage{
GWC(Mw, Ms)
}
\arguments{
\item{Mw}{mass of moisture (g)}

\item{Ms}{mass of oven dry soil (g)}
}
\value{
gravimetric water content (%)
}
\description{
Calculates the gravimetric water content (%)
}
\examples{
temp1 <- GWC(Mw =23,Ms =155);
temp2 <- GWC(Mw = c(25,36,29),Ms =c(155,168,158));
}
