% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{MWD}
\alias{MWD}
\title{Mean weight diameter (MWD)}
\usage{
MWD(WC, ADC)
}
\arguments{
\item{WC}{Weight of soil retained (g)}

\item{ADC}{average diameter between two sieves (mm)}
}
\value{
Mean weight diameter (MWD), mm
#' @examples
temp1 <- MWD(WC = c(4.237,42.42,36.326,20.34),
ADC=c(5,1.1265,0.153,0.0265));
}
\description{
Calculates the Mean weight diameter (MWD)
}
