% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{Particle_Percentage}
\alias{Particle_Percentage}
\title{percentage of sand or clay}
\usage{
Particle_Percentage(RL, RH, MS)
}
\arguments{
\item{RL}{hydrometer reading in calgon solution}

\item{RH}{hydrometer reading in soil suspension}

\item{MS}{weight of oven dry soil (g)}
}
\value{
percentage of sand or clay
}
\description{
Calculates the percentages of sand, silt, and clay
}
\examples{
temp1 <- Particle_Percentage (RL =5,RH =25, MS = 35 );
temp2 <- Particle_Percentage(RL = c(5,4,3),RH =c(25,24,25), MS = c(35,34,36));
}
