% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{RLD}
\alias{RLD}
\title{Relative root density (RLD) (cm/cm3)}
\usage{
RLD(trl, vsc)
}
\arguments{
\item{trl}{Total root length in a core (cm)}

\item{vsc}{Volume of the core (cm3)}
}
\value{
Relative root density (RLD) (cm/cm3)
}
\description{
Calculates the Relative root density (RLD) (cm/cm3)
}
\examples{
temp1 <- RLD(trl=2400,vsc=1507);
}
