% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{RSD}
\alias{RSD}
\title{Root surface density (RSD) (cm2/cm3)}
\usage{
RSD(trsa, vsc)
}
\arguments{
\item{trsa}{Total root surface area in a core (cm2)}

\item{vsc}{Volume of the core (cm3)}
}
\value{
Root surface density (RSD) (cm2/cm3)
}
\description{
Calculates the Root surface density (RSD) (cm2/cm3)
}
\examples{
temp1 <- RSD(trsa=247,vsc=1507);
}
