% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{RWD}
\alias{RWD}
\title{Relative weight density (RWD) (mg/cm3)}
\usage{
RWD(trdw, vsc)
}
\arguments{
\item{trdw}{Total root dry weight in a core (mg)}

\item{vsc}{Volume of the core (cm3)}
}
\value{
Relative weight density (RWD) (mg/cm3)
}
\description{
Calculates the Relative weight density (RWD) (mg/cm3)
}
\examples{
temp1 <- RWD(trdw=250,vsc=1507);
}
