% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{SBC}
\alias{SBC}
\title{Soil binding capacity (SBC) (kg/cm2/plant)}
\usage{
SBC(trdw, r)
}
\arguments{
\item{trdw}{Total root dry weight in a core (mg)}

\item{r}{radius of root (mm)}
}
\value{
Soil binding capacity (SBC) (kg/cm2/plant)
}
\description{
Calculates the Soil binding capacity (SBC) (kg/cm2/plant)
}
\examples{
temp1 <- SBC(trdw=250,r=0.165);
}
