% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{SBD}
\alias{SBD}
\title{Soil Bulk Density}
\usage{
SBD(MS, VT)
}
\arguments{
\item{MS}{weight of oven dry soil (Mg)}

\item{VT}{total volume (m3)}
}
\value{
bulk density of soil (Mg/m3)
}
\description{
Calculates the bulk density of soil
}
\examples{
temp1 <- SBD(MS =155,VT =117.81);
temp2 <- SBD(MS = c(155,149,141),VT =c(117.81,117.81,117.81));
}
