% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{SOCS}
\alias{SOCS}
\title{Soil organic carbon stock (SOCS) (Mg/ha)}
\usage{
SOCS(ar, soc, bd, sd)
}
\arguments{
\item{ar}{area (m2)}

\item{soc}{soil organic carbon (%)}

\item{bd}{bulk density (Mg/m3)}

\item{sd}{soil depth (cm)}
}
\value{
Soil organic carbon stock (SOCS) (Mg/ha)
}
\description{
Calculates the Soil organic carbon stock (SOCS) (Mg/ha)
}
\examples{
temp1 <- SOCS(ar=10000,soc=0.69,bd=1.56,sd=15);
}
