% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{SOCSR}
\alias{SOCSR}
\title{Soil organic carbon sequestration rate  (SOCSR) (Mg/ha/year)}
\usage{
SOCSR(FSOCS, ISOCS, YT)
}
\arguments{
\item{FSOCS}{Final SOC stock (Mg/ha)}

\item{ISOCS}{Initial SOC stock (Mg/ha)}

\item{YT}{Time (Years)}
}
\value{
Soil organic carbon sequestration rate  (SOCSR) (Mg/ha/year)
}
\description{
Calculates the Soil organic carbon sequestration rate  (SOCS) (Mg/ha/year)
}
\examples{
temp1 <- SOCSR(FSOCS=13.62,ISOCS=10.52,YT = 5);
}
