% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{Soil_Indices}
\alias{Soil_Indices}
\title{Calculating the indices from the excel file}
\usage{
Soil_Indices(...)
}
\arguments{
\item{file_path}{path for your excel file}
}
\value{
different weed indices
}
\description{
Calculates different weed indices from the given excel file
}
\examples{
file_path <- "G:/Naveen_spectral_similarity_index/Weed_Indices/Weed index-1.xlsx";
all_Weed_indices <- Soil_Indices(file_path);
}
