% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{SPD}
\alias{SPD}
\title{Soil Particle Density}
\usage{
SPD(MS, VS)
}
\arguments{
\item{MS}{weight of oven dry soil (Mg)}

\item{VS}{volume of solids (m3)}
}
\value{
particle density of soil (Mg/m3)
}
\description{
Calculates the particle density of soil
}
\examples{
temp1 <- SPD(MS =155,VS =59.05);
temp2 <- SPD(MS = c(155,149,141),VS =c(61.52,59.05,53.26));
}
