% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{TP}
\alias{TP}
\title{Total Porosity}
\usage{
TP(BD, PD)
}
\arguments{
\item{BD}{bulk density (Mg/m3)}

\item{PD}{particle density (Mg/m3)}
}
\value{
Total Porosity (%)
}
\description{
Calculates the total porosity (%)
}
\examples{
temp1 <- TP(BD =1.53,PD =2.65);
temp2 <- TP(BD = c(1.55,1.49,1.41),PD =c(2.61,2.59,2.53));
}
