% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{VR}
\alias{VR}
\title{Void ratio (VR)}
\usage{
VR(tp)
}
\arguments{
\item{tp}{total porosity (%)}
}
\value{
Void ratio (VR)
}
\description{
Calculates the Void ratio (VR)
}
\examples{
temp1 <- VR(tp=49.5);
temp2 <- VR(tp =c(48,46,49));
}
