% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{VWC}
\alias{VWC}
\title{Volumetric water content}
\usage{
VWC(gwc, bd)
}
\arguments{
\item{gwc}{gravimetric water content (%)}

\item{bd}{bulk density of soil (Mg/m3)}
}
\value{
volumetric water content (% or cm3/cm3)
}
\description{
Calculates the volumetric water content (% or cm3/cm3)
}
\examples{
temp1 <- VWC(gwc =23,bd =1.55);
temp2 <- VWC(gwc = c(25,36,29),bd =c(1.55,1.68,1.58));
}
