% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgroSoil.R
\name{WFPS}
\alias{WFPS}
\title{Water filled pore space (WFPS)}
\usage{
WFPS(gwc, bd, pd)
}
\arguments{
\item{gwc}{gravimetric water content}

\item{bd}{bulk density (Mg/m3)}

\item{pd}{particle density (Mg/m3)}
}
\value{
Water filled pore space (WFPS) (
}
\description{
Calculates the Water filled pore space (WFPS)
}
\examples{
temp1 <- WFPS(gwc= 13,bd=1.31, pd=2.53);
temp2 <- WFPS(gwc = c(14,16,19),bd =c(1.35,1.19,1.23),pd=c(2.54,2.45,2.31));
}
